+++ xbmc/pvr/utils/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(SOURCES PVRPathUtils.cpp)
+
+set(HEADERS PVRPathUtils.h)
+
+core_add_library(pvr_utils)
